requirements.txt
asyncio==3.4.3
fastapi==0.68.0
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.0
starlette==0.14.2
uvicorn==0.14.0
